requirements.txt
# I've pinned these to what Worked On My Machine™ but feel free to upgrade
flask-slack-events==0.0.2
gunicorn==20.0.4
flask==1.1.1
requests==2.22.0
sentry-sdk[flask]==0.13.5
